requirements.txt
langchain-core~=0.3.65
langchain-community~=0.3.25
langchain-ollama~=0.3.3
langchain-tavily~=0.2.2
langserve[all]~=0.3.1

chromadb~=1.0.12